Help user when package has no built versions
Ask user which version they are looking for and redirect them to the
appropriate page.

First I thought about simply explaining which URL the user has to go to
but this seems friendlier.
My second idea was to do the redirect in JS on the client directly, but
I decided that the extract work is worth it to not depend on JS.

redirect-to-version is similar to jump-to-project. Maybe we can find
some way to combine them?

Also I reused the validation of request-build. Not sure if that's the
right way.