Bold artifact coordinates for request-build-page (#179)
This makes it clearer which parts cljdoc is using for searching and can
help show people why they might have gone wrong. Adding the "v" in front
of the version by itself looks slightly confusing. It's not obvious that
it was added in the formatting, but wasn't included in the search.