Refactor sidebar rendering
There now is one function to render the sidebar. This function receives
the same arguments as the other main rendering entry points:
`route-params` and `cache-bundle`.

The small notes indicating a problem have been updated to more
accurately report what is going on, specifically the following scenarios
are handled:

- No articles were found at all.  
  Causes: No Git repo found or unsupported document types. 
  Solution: Point user towards relevant documentation.
- Only Readme and Changelog were found.  
  Causes: Articles haven’t been set up.  
  Solution: Point user to articles documentation
- No namespaces were found. 
  Causes: Failed analysis or Uber-module artifact without actual sources. 
  Solution: Invite users to open an issue.

The separators have also been changed slightly so that the separator for
the Articles section is only shown if there is a Readme and/or Changelog
article.