Search: fix no match for "<group> <artifact>"
e.g. "metosin muuntaja". The multi-tokens->query
required that all tokens are present (MUST) but the problem is
that it is run for each field while the tokens might be in different
fields.

We still need to do something so that hits with _all_ the tokens
(possibly across multiple fields) are the top.