Search: Require presence of all tokens in a longer query
The key change is that "SHOULD" is replaced with "MUST" regarding of the
occurrence of the search tokens, i.e. they all have to appear in a "document"
for it to match. Thus "metosin muunt" will only yield results that have both,
not those that have either (with the actual metosin/muuntaja way down the list).