upgrade asciidoctorj library
Adapt code to new asciidoctor factory method signature.

This moves us to JRuby 9.x which is probably good since the JRuby 1.7
end-of-life was in May of 2017.

This is a big change JRuby-wise so we should monitor for such things
as performance hits.

Note: previously we were bringing in jcommander and ruby deps to overcome
some tooling problem, this is no longer necessary.