add clj-kondo linting
Overview
--------
Closes #332

- Added clj-kondo with appropriate config to lint cljdoc sources.
- For this first run, found mostly unused namespaces and bindings.
- Corrected or suppressed all errors/warnings produced by clj-kondo.
- Now failing build on circleci should any new errors/warnings crop up.

Run the new `script/lint.sh` to lint cljdoc sources.

Please continue reading.

Unused destructured keys
------------------------
One pattern in cljdoc is the employment of unused destructured keys as means
of documenting what is expected for a function. For example, given:

`(defn [{:keys [a b c d] :as e}])`

cljdoc might not use all of `a` `b` `c` `d` and `e` in the function body.

I've deferred the decision of whether or not cljdoc team wants to continue with
this practice via script/lint.sh which suppresses this specific case of unused
binding.

Issues
------
The following should be reviewed before this commit is merged.

**unlogged exceptions**

We have a couple of cases of caught exceptions not being logged. They should be
reviewed and a comment added if this is intentional.

See:
- [server/api.clj](https://github.com/cljdoc/cljdoc/blob/d54cb58b20d2b4cb1e7c714b3637aa8b89956ce0/src/cljdoc/server/api.clj#L38)
- [server/pedestal.clj](https://github.com/cljdoc/cljdoc/blob/d54cb58b20d2b4cb1e7c714b3637aa8b89956ce0/src/cljdoc/server/pedestal.clj#L295)

**empty code**

An unused function with a big comment and an empty body had me puzzled. Was
I right to delete it?

See:
- [git_repo.clj](https://github.com/cljdoc/cljdoc/blob/d54cb58b20d2b4cb1e7c714b3637aa8b89956ce0/src/cljdoc/git_repo.clj#L219)

**oversights?**

The comment above an unused binding made me think there might be a
programming error - as such I left the code in and temporarily suppressed
the linter warning via an underscore prefix.

See:
- [doc_tree.clj](https://github.com/cljdoc/cljdoc/blob/d54cb58b20d2b4cb1e7c714b3637aa8b89956ce0/src/cljdoc/doc_tree.clj#L89)