downgrade pedestal to 0.5.3
There has been a breaking change in version 0.5.4 that we seem to
have missed when upgrading. That change affects how the Pedestal
service will bind to the network and currently makes cljdoc fail
to properly run inside Docker.

Solution isn't yet fully clear to me so this is a quickfix for now