Use full file path for exists checks (#378)
Due to Clojars recently moving to S3 as a backing store for the
repository CDN at https://repo.clojars.org/, checking for the
existence of a directory for a group/id/version/ will no longer work
to confirm the existence of a version - non-existent paths will still
return a 200 with an empty index file.

This changes the existence logic to instead look for
group/id/version/maven-metadata.xml for SNAPSHOT versions, and
group/id/version/id-version.pom for non-SNAPSHOT versions. Those paths
*will* return 404 if the version does not exist, and this check should
work for any maven repository.

This also includes a little cleanup to use the new jar/pom-uri &
snapshot-version? helper fns.