Allow library revision to be forced when testing locally
The cljdoc cli ingest command has a --rev option to support
local testing.

When a specific --rev is requested, we are now disregarding
any git tag representing the project version. This project
version tag is preferred at render time - something that we
don't want when we are forcing a --rev.