Update running cljdoc locally docs
After some exploration (see ADR 0020), I realized that a documentation
change was sufficient to allow for a full library doc preview when
testing locally.

AsciiDoctor turned out to be a better format for the updates, so
switched to that, and updated all references from old .md version.

Old guidance, which encouraged an import from the local file system
only, has been preserved as a viable quick verification.

Notable additions:
- full local preview - which imports from hosted SCM rather than filesystem
- ingest from published library - for the cljdoc developer who has no
real interest in testing against a local library
- overview diagram of how testing locally works.
- now showing both docker and cljdoc source commands
- added section on verifying offline bundles

Switched example library to cljdoc-exerciser - and made it easy to
switch to a different example project should we want to.