Prevent default behavior for Ctrl-K
On Firefox and Chrome (the ones I tested), when you press Ctrl+K
the browser's navigation bar gets focused and adds a question mark
for searching, which prevents the user from using this as an actual
feature. Calling `preventDefault` on the event, makes the cljdoc
search box to get focused instead.

Reference:
  * Chrome: https://support.google.com/chrome/answer/157179
  * Firefox: https://support.mozilla.org/en-US/kb/keyboard-shortcuts-perform-firefox-tasks-quickly#w_search