Tweak calls to badge service
Because badgen.net will often (always?) return an error on the first
call to a badge it does not have in its cache, we introduce 1 retry
on failure after a 300ms sleep.

Also we now return a 503 to the client if we can't manage to get
a sucessful return from badgen.net. We were previously returning 200
on failure with a request to cache for 30 minutes, so a browser refresh
would not always retry to fetch a failed badge.

Fixes #422