Update npm deps - prettier
We use prettier to check and format JavaScript sources.

I added a .prettierrc.json to override some new defaults that prettier
introduced. Some minor formatting changes were still applied, but they
all seem reasonable.

There did not seem to be a need for prettier-check anymore as this
functionality is available in prettier itself.