Add retries to maven fetches.
The artifact indexer was failing for me every time it ran. The maven
servers kept giving us back 429 responses which mean we should back
off on our request rate a little.

So for the implementation I went to add retries with backoff and
discovered that we're using io/reader with a url to fetch from maven.
Since there's no way to get a response map and handle this another way
we need to catch the exception and retry the operation.

There are libraries we could pull in to catch exceptions and retry
operations but they're all a little overcomplicated in their own way.
Given how simple of a need we have it seemed a good idea to write it
ourselves. If our needs grow then it's probably a good idea to pull
in a library instead.