Retry maven fetches with clj-http-lite.
The artifact indexer was failing for me every time it ran. The maven
servers kept giving us back 429 responses which mean we should back
off on our request rate a little.

So for the implementation I went to add retries with backoff and
discovered that we're using with-open with io/reader on a url to pull
docs from maven. This is less than ideal and we have clj-http-lite
already in the project so it makes sense to switch to that.

This change moves to clj-http-lite and adds robert-bruce for retrying
with backoff on exceptions when requests fail. Now download and
indexing doesn't fall over!