First pass at adding an API endpoint for docsets.
We need to feed the docset search with documentation in JSON format
to feed into the client-side search engine. We re-use the cache-bundle
here as the source of this information. We will likely keep on using
this but trim down the data to something more reasonable before sending
it back.