Add previously visited doc links on homepage II #282
There's already an abandoned PR for this issue #399 so I thought I'd give it another try. I kept only the first commit of the old PR since merging the complete old PR  would be more effort.

After the page is loaded existing example links are replaced with the most recently visited doc links. If one of the initial example links is used it get's ignored.