Sanitize HTML converted from library markdown
Documentation of libraries that cljdoc imports include markdown in the
form of:
- AsciiDoc and CommonMark articles
- CommonMark in docstrings

We now sanitize this markdown after it is converted to HTML to:
1. primarily protect cljdoc users from potentially malicious HTML
2. secondarily ensure separation of user markdown styling from cljdoc
   styling

Testing has been done, but if we learn that we are over-sanitizing,
this is easy enough to correct.

See cljdoc.render.sanitize namespace for more verbiage.