Only render [[wikilinks]] when appropriate
We now only render wikilinks as links when:
- wikilink rendering is enabled (i.e. for docstrings only)
- when the text inside the wikilink resolves to an existing namespace
  or var in the library
- wikilink syntax does not include text, ex. [[my.ns/var|text]] will
  render as self

Although we make use of elements of the flexmark wikilink extension
we no longer register it as an extension.

One could argue that we should just write our own flexmark extension
for cljdoc wikilinks, but this seems to be working for the time being.

Closes #507