Rework library articles tips
Formerly we were suggesting to the reader that they might want
to glance at our Articles docs. This messaging always appeared when a
library did not have more than readme and/or changelog articles.

This did not make a ton of sense and was a bit noisy/wordy.
As raised in the issue, we now assume the library author does not need
a tip if we have found their project's cljdoc.edn config.

We instead now offer a click-able "tip" in the articles header.
It is noticeable but also subtle.
When clicked, the tip shows tip content to the user.

The tip appears when:
- there is no cljdoc.edn and we've auto-discovered articles
- there are no articles because scm cannot be found
- there are no articles in a found scm

To make this work, we've move all articles under the articles header.
Formerly readme and changelog appeared above the articles header.

The new tip more effectively serves our 2 audiences:
- the library author who might want to make a change/fix
- the doc reader who will usually ignore our tip, but if curious, can
learn more.

Also: fixed vertical spacing on article toc for tocs with hierarchy.

Closes #270