Allow override for articles git revision (#675)
Cljdoc now looks at the new `cljdoc-<version>` git tag as the revision
to both import articles at build time, and resolve references from
articles at render time (such as images).

Added a cljdoc/cljdoc-test-repo to support git analysis testing.
Not sure if this is the best idea, but it is what we are doing for now.

Of note:
- Removed looseness around finding articles. If a cljdoc.edn specified
article was not found at the expected revision a version from master
HEAD was used. My feeling is that being helpful in this way can be more
confusing and sometimes not helpful. And with this new feature, folks
can adjust after the fact.
- Includes minor refactor of derived toc, it now slurps articles the
same way that process toc does.
- Deleted unused vars when I noticed them, made local vars
private (thanks clojure-lsp!)
- Trailing whitespace on inferred doc tree titles now stripped.

See author guide for details.

Closes #31