maint ci: use current clojure bb, yaml aliases
I am getting:
class java.util.HashMap$Node cannot be cast to class
java.util.HashMap$TreeNode

when downloading deps via Clojure CLI.
This is supposed to be a fixed issue, so I'll to move to the
current release of Clojure to and see if that fixes things.

Also, CircleCI image includes babashka, but a stale version,
so let's use latest and greatest.

Also, I've learned that we can use YAML aliases with CircleCI
configuration, so I've incorporated those to make our config
a little less copy/paste.