ci: deploy: retry check for cljdoc docker tag
It can take a while for a pushed docker image to become visible on docker hub.
I have noticed sporadic failures in the past.

On 2022-10-17 I was seeing repeated failures, one case taking 20m for a pushed
tag to appear. I assume this was just a bad day for docker hub.

Start with what seems to be reasonable retry timeout of 3m with 2s interval.

includes: rework of wait-until fn to do more logging