maint: bump deploy module deps
While diagnosing a deploy issue, I stumbled on some ancient deps for our
deploy module. I switched from clj-http to clj-http-lite (because we use
that in cljdoc itself, and we have not need for the beefier clj-http).

It all seems good, but will tweak based on success (or lack thereof) of
next deploy.

Also added the deploy module to our `bb outdated` check.