Handle case: listed artifact does not resolve
I have learned that that artifacts listed by clojars sometimes do not
resolve. One example is an artifact with a version of
just "SNAPSHOT".

The cljdoc release monitor pings clojars for new releases and will
endlessly retry these unresolvable, but listed, artifacts.

There are not many cases of these, but they do generate exceptions which
is confusing noise for the cljdoc devops team.

There are other flows that trigger doc builds, this change is limited
to the release monitor flow. This is the one generating the logging
noise.

The release monitor now does a pre-check.
The pre-check tries to resolve the artifact, if resolve returns
- 200 - good to attempt a build
- 404 - create and immediately fail build with listed-artifact-not-found
- else - requeue release, if after 10 tries, still no go, create and
immediately fail build with listed-artifact-resolve-failed.

Once a build is created it is effectively no longer in the queue.
Queued builds with fewer retries are favored.

Related changes of note:
- build log adjusted to key off analysis_requested_ts instead of
analysis_triggered_ts, not sure why this was not already the case.
It allows for our new pre-check failures where analysis_triggered_ts is
not populated.
- recent builds query was having sqlite convert 'now' to a localdate.
Our dates are in utc so we don't want to do that. Was harmless on prod where
localdate is utc, but not great for dev where localdate is dev box specific.
- release monitor checks to clojars refined slightly. Was previously
adding 1 second to last known release timestamp, now simply excluding the
first release (because it was in the last release fetch).

Closes #716, closes #717