When library not found present 404 page
Previously we only presented a page not found 404 page when a route was
not found.

When a library was not we'd throw and generate a 500 error.

A library not found error happens when a user directly manipulates or
scripts a cljdoc URL. You'd think this would be rare, but our logs tells
us it happens enough to be worthy of addressing.

This change adjusts our existing 404 page to:
- provide a link back to the cljdoc home page (was missing)
- include some details of the library that was not found

Impacts:
- adjusted client side visit history/switcher to ignore library 404 errors
- adjusted pom db caching (in cache.db) to never cache pom that is not
found (we were previously throwing on pom not found). Adjusted tests
accordingly.

Closes #720