Upgrade to highlight.js v11
I was on the fence about this one as v11 removed auto-merging of HTML
into formatted blocks. This feature supports AsciiDoc callouts.
See: https://docs.asciidoctor.org/asciidoc/latest/verbatim/callouts/ if
you need a reminder as to what these are.

After more carefully reading over highlightjs/highlight.js#2889, I
realized maybe it wouldn't be very difficult to simply re-introduce the
feature. The issue even includes a complete JavaScript code example to
do so.

I did my naive best to transcribe the sample JavaScript to TypeScript
(which even caught a couple of minor issues).

It took me some time to figure out a way to expose the TypeScript
function to our layout page. Parcel.js used to provide a --global
option to support this use case, but that feature was removed.

In the end I just assigned the function to the browser window to make it
available.

There is a new console warning in highlight.js v11 that is generated
when formatted code contains unescaped HTML. I turned this off because
the adoc callouts add... you guessed it... unescaped HTML. We also have
good HTML sanitization, so I think we are good here.