build log: don't attempt to truncate to short sha
The build log was trying to distinguish a version tag vs a sha based on
length only and then truncating a sha to a short-sha.

This is too naive and is truncating version tags longer than 8 chars.

For now we'll just blurp out the whole version/sha.
To me this looks fine and is acceptable for a low traffic web page.

If we ever want to revisit, we'll have cljdoc store the necessary info in the
build log so the render code does not have to make naive guesses.

Closes #757