Fixups for sidebar scroll pos save/restore
Switched from localStorage to sessionStorage for scroll state.
I think being more ephemeral here is the better way to go.

Noticed that restored scroll position was a bit off every time.
Waiting for the dom to load fixed this jitter.

Now saving scroll state when a user clicks on a sidebar item.
If that state is present and is for the same doc lib/version on load, then
we restore the scroll pos.

Otherwise, if the selected sidebar item is out of view, we scroll to it.
This supports things like bookmarks.

Noticed that style for highlighting of article and api items were different.
They are now both a tachyons `b`.

Because our docset search input is effectively part of the sidebar and scrolls
with the sidebar, I opted to not to do any automatic scrolling when the user
selects a docset search match.

Closes #790