Support punctuation in single docset search terms
Searching a single docset is a client-side feature.
It now supports searching for text that contains punctuation.

This can be helpful when searching for vars and/or namespaces which can
contain punctuation.

Highlights of changes:
- strongly favour exact match of terms with punctuation but also match
on subterms. For example a search for `cloure.core.test` would strongly
prefer an exact match but also match on terms `clojure`, `core` and
`test`.
- turn off stemming support, I think we are ok without it. We already
do expansion matching (i.e. `ab` matches `abc` `abracadabra`, etc), and
stemming might be less helpful for source code elements.
- turn off stop words. Stop words are normally interesting for prose,
but for code, we probably don't want to exclude common words.
- turn off trimmer, we handle trimming ourselves in our new tokenizer

Closes #813