Migrate from clojure.java.jdbc to next.jdbc
The intent is the same behaviour as before.

To avoid too much rework, asked next.jdbc to return clojure.java.jdbc
compatible result sets with `:builder-fn rs/as-unqualified-maps` option.

Opted to not ask next.jdbc to use JDBC's `Statement` `getGeneratedKeys`.
We were transparently taking advantage of this in clojure.java.jdbc.
But this method was deprecated by sqlite-jdbc because it is unreliable.
It was removed in v3.43.0.0 (which we'll upgrade to separately).

We instead rely on SQLite's RETURNING syntax to return generated keys.
This was only needed when inserting a new build log entry.

Next.jdbc does not have funtions to abstract the creation of tables,
so simply generated the necessary SQL to create our cache table.

I took advantage of next.jdbc sql sugar syntaxes when the change
was small.

Closes #827