lint: tweak clojure vs bb deps
Use new "bb print-deps" for full list of bb dependencies including
built-ins. Tack this onto clojure deps when calculating classpath for
lint cache build.

To support Windows, with its limited command-line length, move to
building cache programmatically in clojure in build.clj.

This is not perfect, if a bb dep version is older than a clojure dep version,
we'll end up using the older dep, but it is better than before and unlikely to
cause issues.