docker: increase max jvm heap size
Our docker jvm max heap is reported as:

```
$ nomad alloc logs -stderr 488e4d6d
VM settings:
    Max. Heap Size (Estimated): 386.69M
    Using VM: OpenJDK 64-Bit Server VM
```

Our current docker container is 1600mb.
The default ram percentage used by the jvm is 25% or ~400mb.
This is a conservative default.

As far as I remember, we don't have much else going on in our cljdoc
docker container. So let's try bumping the ram percentage to 75% or ~1200mb.