ci: address machine deprecation warning
CircleCI now prefers we use "default" instead of specific machine image tags.

This means we might need to adapt when CircleCI updates the underlying
machine image but... makes us a better CircleCI citizen? And means we'll
not need to periodically update this config.

Also addressed security warning about docker login and moved login step
to where it is actually needed.

Closes #871