Get traefik working
That was tough! So many 404s until I finally realized that traefik now
expects consul discovered services to be configured by nomad tags.

I think I have an equivalent config to cljdoc production.
(We'll enable real https certs when we activate domains for Exoscale)

I moved consul back to running as root (no different than it is
currently in production).
And bound it to 0.0.0.0.
These might be extreme.
But might also be ok.

Consul's config is now consistent with nomad's in that is now in a config
file instead of on the command line.

I switched to a maintained fork of jsch to get support for modern
ciphers like ed25519.