Switch from aws-api to aws sdk v2 via interop
Bit of a bummer, but aws-api was copying entire object to heap, and we
don't have enough heap for that. Our database backup is close to 1gb.

Abstracted s3 to its own namespace and protocol to make swapping aws-api back
in the future. Maybe they'll fix the issue, or maybe I was just using it wrong.

See: https://github.com/cognitect-labs/aws-api/issues/257