ops: give a wee bit less RAM to cljdoc heap
Nomad is noticing that RAM is exhausted on cljdoc alloc and killing it.
This resulted in a brief outage (as reported by UptimeRobot).

(Cljdoc is not noticing any problems at all and merrily walks on the
edge of resource exhaustion).

We currently tell nomad to give cljdoc 1600mb.
We tell our JVM to use a max of 75% of this.
(It estimates a max heap of 1.13gb on startup which is actually about 70%?)

We'll tell the JVM to use a max of 70% of container RAM and see if that
helps matters.