ops: support local testing of deploy
This change supports testing deploys against a local host VM.

Parameterized cljdoc profile in the nomad job spec.
(It was hardcoded to prod which is not appropriate for local testing).

Traefik's Let's Encrypt endpoint is now set the staging endpoint unless
cljdoc profile is prod (we don't want to be requesting real SSL certs
unless in production).

Move traefik config from toml to edn to support easy parametization.
And config is now sent as yaml to consul->traefik.
Traefik also understands yaml and it is easy for us to convert edn to yaml.