ops: support visualvm -> cljdoc server
This was a confusing slog!

Visualvm talks to jmx over rmi.
The RMI handshake starts off with a request to an RMI registry that returns
the hostname on which to continue the session.

A docker container deployed by nomad is bound to the host IP.
I haven't found a way around that. It thing that for normal nomad usage this
is probably reasonable, our single node usage of nomad is unconventional.

The only way I can see to expose a nomad docker deployed cljdoc server port is
through traefik. This is a bit non-intuitive to me because I saw traefik
as our public front end, not a private ops front-end.

Traefik does support routing tcp traffic.
Docs, as usual, are on the brief side.
But I stumbled on the correct incantation.

So what we have is:
- cljdoc server jvm process enables jmx on port 9010 with an rmi hostname of
localhost. We must set both the port and rmi.port to the same port,
otherwise the connection will be attempted over several(?) ports
(see script/docker_entrypoint.clj jvm properties on launch).
- nomad exposes internal docker port 9010 to a dynamic jmx port bound to the host server
IP. We must use a dynamic port because we need to support blue/green
deploys and a static port would mean a port collision during deploy.
- traefik is configured to expose port 9010 and route that to our nomad
dynamic jmx port.

When we ssh port forward in, we satisfy the RMI handshake, the registry
returns localhost:9010 which goes through:
- traefik port 9010
- which traefik maps to our dynamic jmx port on the cljdoc docker container
- which nomad had docker map to port 9010 inside the docker container
- which is received by the jvm

Port 9010 is not exposed by our firewall, so although it is exposed by
traefik, it is private and ops-team-only via ssh port forwarding.