Reuse Lucene IndexReader
The Lucene IndexReader is threadsafe and the Lucene docs recommend that
it be reused. I never bothered doing this because it does add some
stateful complexity in that the IndexReader needs to be recreated if
necessary (this will happen if the Lucene index has changed).

On prod we are seeing our RAM usage creep up slowly until it hovers
around our docker container limit. This RAM usage is outside of the JVM
heap.

The RAM usage might be Lucene's use of memory mapped files.
I was curious to see if using Lucene's IndexReader in the recommended
way would impact RAM usage.

The IndexWriter can also be long-lived, but our writes are so infrequent
I do not currently see the benefit to keeping an IndexWriter around.
And not closing an IndexWriter can lead to Lucene index corruption, a
scenario I'd like to avoid.