Add caching for maven central artifacts
In the spirit of being good citizens, reduce our request to maven
central search API.

Caching means:
- requests for artifacts for group will only occur if there are new
artifacts for the group
- requests for artifact description happens only for artifacts
that have new releases

Keep it simple:
Cache our 3 groups to edn which we store as as resources.
This cache can be committed whenever a developer notices, which should
happen naturally during local dev.