ops: bump traefik
Martin received a "certificate expire notice" email from letsencrypt.
If we look at cljdoc ssl certs they are slated to expire on Feb 6.
Which matches the email.

Does it mean that traefik is not auto-renewing our certs?
I don't see anything in the traefik.log.

If I read https://letsencrypt.org/docs/expiration-emails/, maybe this is
nothing to worry about?

Let's bump traefik to see if that has any effect.