Review memory usage around ls-files
In (I think) an attempt to be composable, cljdoc was separating getting a
list of git files from computing the sha256 of those files.

This had a negative side effect of consuming lots of memory for large
git repos during git repo analysis. For example, for
https://github.com/coconutpalm/clojure-desktop-toolkit this consumes
1.2gb which results in an out of memory error on our frugally deployed
production server.

I've replaced `ls-files` and `path-sha-pairs` with a single
`sha-file-map`. For the big clojure-desktop-toolkit, this results in a
much more modest 746kb of memory usage.

See #1027