dev: try out borkdude's new unused-deps
Very cool. Current output is:

```
[ TASK unused-deps ]-------------------------------------------------------------
Dependencies that do not seem to be referenced by code
NOTE: some unused deps are overrides to bring in later version of a dep to compensate for bugs or CVEs

[ch.qos.logback/logback-classic #:mvn{:version "1.5.18"}]
[dev.weavejester/ragtime #:mvn{:version "0.9.5"}]
[org.apache.lucene/lucene-queryparser #:mvn{:version "10.2.1"}]
[org.asciidoctor/asciidoctorj #:mvn{:version "3.0.0"}]
[org.eclipse.jetty/jetty-alpn-client #:mvn{:version "11.0.25"}]
[org.eclipse.jetty/jetty-alpn-server #:mvn{:version "11.0.25"}]
[org.eclipse.jetty/jetty-servlet #:mvn{:version "11.0.25"}]
[org.eclipse.jetty.http2/http2-server #:mvn{:version "11.0.25"}]
[org.eclipse.jetty.websocket/websocket-jakarta-server #:mvn{:version "11.0.25"}]

TASK unused-deps done.
```

The jetty deps are CVE overrides.
Not sure about lucene-queryparser.
asciidoctorj and ragtime: false positives?