fix scroll indicator for protocol members
The scroll indicator (shown as `|`) was sometimes showing like so:

```
| ProtoOne
   member1
   member2
|  member3
|  member4
```

This might be technically correct, we are still in the ProtoOne block
when member3 and member4 are visible (and member1 and member2 are not visibile).
But I found the gap confusing from a usability perspective.

So now we'll just show:
```
  ProtoOne
   member1
   member2
|  member3
|  member4
```