TIL: cljdoc badge error handling
Cljdoc badge will error if LATEST available artifact docs not built.
I was thinking maybe badge would show latest version built instead.
It has worked this way for ages, so maybe what we want?

Adjusted badge image error to clearly distinguish between a library not
being found vs docs not built for latest library version.

A difference before was that we returned 404 in on not found (not
an image). This would have shown as a broken image.