maint: sqlite impl
Turf HugSQL. It was used because variadic clauses in JDBC have no
special support. But isn't rocket science to whip up the appropriate
queries at the JDBC level. So did that.

Turf Tufte. It might be awesome, but I've not used it during my time
with cljdoc. So it is gone. We can always bring it back in the future
if needed/wanted.

Reworked `bundle-docs`. Take the special and rare case of loading docs
for submodules out of the normal path. This makes the normal path much
more obvious and also avoids unnecessary queries when there are no
submodules specified by a library.

Closes #1110