Adapt to changes in maven central repo
`etag` no longer present on dirs, but is still present on files.
Now saving `last-modified` and `etag` to our cache.
Preferring `etag` when present, but now falling back to `last-modified`.

We are now trusting `last-modified` is updated when any child is updated.
Not 100% confident this is true. We'll adapt if we discover our artifact
group caches are stale.

Observation: Maven Central has switched to weak etags (notice the
`"W/\"some-etag-here\""` in the edn cache files)

Closes #1121